[IC] Multi-module incremental compilation smoke tests for KMP
Part 2. Main multi-module KMP smoke tests for incremental compilation

^KT-56963 Fixed
^KT-63876 Fixed

Merge-request: KT-MR-13193
Merged-by: Evgenii Mazhukin <evgenii.mazhukin@jetbrains.com>

+++ libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/resources/testProject/generic-kmp-app-plus-lib-with-tests/lib/src/jsTest/kotlin/TestDependsOnCommonMainAndPlatformMain.kt
@@ -5,7 +5,7 @@ class LibJsTest {
 
     @Test
     fun jsUtilReturns0() {
-        assertEquals(0, libJsPlatformUtil())
+        assertEquals(0, libJsPlatformUtil().toInt())
     }
 
     @Test
