+++ libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/resources/testProject/generic-kmp-app-plus-lib-with-tests/lib/src/jvmTest/kotlin/TestDependsOnCommonMainAndPlatformMain.kt
@@ -5,7 +5,7 @@ class LibJvmTest {
 
     @Test
     fun jvmUtilReturns400() {
-        assertEquals(400, libJvmPlatformUtil())
+        assertEquals(400, libJvmPlatformUtil().toInt())
     }
 
     @Test
