+++ libraries/tools/kotlin-gradle-plugin-integration-tests/src/test/resources/testProject/generic-kmp-app-plus-lib-with-tests/lib/src/nativeTest/kotlin/TestDependsOnCommonMainAndPlatformMain.kt
@@ -5,7 +5,7 @@ class LibNativeTest {
 
     @Test
     fun nativeUtilReturns800() {
-        assertEquals(800, libNativePlatformUtil())
+        assertEquals(800, libNativePlatformUtil().toInt())
     }
 
     @Test
